refactor/feat(codebase): Total refactor and API-compatible. Added many features.
In this refactor we have added new features too:
[x] Created a better organization of files and types.
[x] Created a way of communicate with our children components
[x] An API more robust and open
[x] Created external events to get changes from inputs (onChange, onBlur, onChange, onEnter, onEscape, onFocus, etc.).
    These events are shotcuts to angular-native events (ex. keypress.enter/keypress.escape), so you can bind those. However, this
    shorcuts provide you of state of input when this one changed.
[x] New options to configure when an event is triggered (ex. hidebuttons, cancelOnEscape, saveOnEnter, saveOnBlur, etc.)
[x] An internal service shared between root and child component. (this feature will be more use in the future, so it is WIP)
[x] Interfaces to provide a specifig config to inputs (specific intellisense to input text, input select, etc.)

How works The new API to communicate the root component with their children?
When a child component change its state, this one triggers an event and the root component get it.
Using this system of internal events we can provide external events (to the client) and trigger events such as onChange, onFocus, on Blur, etc.
A change of state can be: change of value, edit mode or disabled mode.

Closes #38 #47 #30 #21