feature(InlineEditor): If an input is empty or not is handler by itself
This fix a problem with `InputSelectComponent`, it showed `empty` when the option was not found, but `empty` property of its state was `false`. It happened because the value was not empty (`"test"`, `1`), but there was not any option with that value.

Now the inputs handle when they are or are not empty implementing the `isEmpty` method of `InputBase`, by default an input is empty if its value is an empty string (`""`) , `undefined`  or `null`.

And I have simplified the way how the `InputSelectComponent` searched the option selected.